Jerry/requirements.txt
langchain
langgraph
langchain-core
langchain-community
langchain-ollama
chromadb
faster-whisper
pyautogui
ddgs
streamlit
SpeechRecognition
openwakeword
requests
psutil
pyperclip
googlesearch-python
pyttsx3
pyaudio
numpy
pillow
